debug hamburger menu button in mobile
