+++ public/index.html
@@ -166,6 +166,7 @@
     <!-- Counters -->
     <script src="js/jquery.countTo.js"></script>
     <!-- Main -->
+    <script src="js/setup.js"></script>
     <script src="js/main.js" type="module"></script>
   </body>
 </html>
